humidity conversion TDD passing

+++ smart-fridge/main/data/handle_humidity.h
@@ -8,7 +8,7 @@ extern "C" {
 #include <math.h>
 #include <inttypes.h>
 
-float bytes_to_humidity(uint8_t humidity_msb, uint8_t humidity_lsb);
+double bytes_to_humidity(uint8_t humidity_msb, uint8_t humidity_lsb);
 
 #ifdef __cplusplus
 }
